moved to new installer + changed icon + ... - only launch one instance of application - hwinfo option now invisible if plugin not found - new installer innosetup instead of nsis (reccomended to uninstall old version first) - changed readme a bit - added a prefixtext to licensefile for new installer - updated hwinfo64.dll - lock shared memory when readinf from aida - about now shows versionnumber

+++ FanControllerV2/HWInfo.h
@@ -1,4 +1,3 @@
-
 /*
 * Project: FanControlV2 host apllication for 6_channel_usb_fancontroller
 * Author: Peter Repukat
@@ -28,7 +27,6 @@
 #include "SettingsHandler.h"
 #include "ui_fan.h"
 
-
 class HWInfo
 {
 private:
@@ -37,13 +35,13 @@ private:
 
 	int sensorValues;
 	QString *tempValues = NULL;
-	
+
 	QListWidget *lW_Status;
 
 	QString TemperatureEntry;
 
 	bool readSucces = false;
-	
+
 	SettingsHandler* Settings;
 
 	ui_fan::AutoPages* AutoPages;
@@ -54,10 +52,9 @@ private:
 
 	char temp[4096];
 
-
 	//hwinfo64 sensor api shouldn't be made publicly.. so i only provided a proprietary dll...
 	//since this program is gpl licensed, i use the plugin as an optinal one and load it dynamically, so it could live without the application
-	typedef void(__cdecl* _readHWiNFOTemps)(char* c_str, int len);
+	typedef void(__cdecl* _readHWiNFOTemps)(char* c_str, size_t len);
 	_readHWiNFOTemps readHWiNFOTemps = NULL;
 
 	bool retry = true;
